FlightTaskAuto: remove _land_heading and instead directly work with _yaw_setpoint
When weather vane is enabled, we don't want to set any heading setpoint during Land,
unless land help (nudging) is enabled and the yaw stick deflected.
This functionality was already there and acting on the _yawspeed_setpoint, but
not yet on the _yaw_setpoint, as that one got overwritten by the _land_heading,
that in turn was set during the init and only overwritten by the nudging (but not WV).

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>